+++ engine/src/main/java/io/zeebe/engine/processor/workflow/handlers/multiinstance/MultiInstanceBodyCompletedHandler.java
@@ -114,7 +114,7 @@ public final class MultiInstanceBodyCompletedHandler extends AbstractMultiInstan
 
     variableReader.wrap(array, 0, array.capacity());
     variableReader.readArrayHeader();
-    variableReader.skipValues(index - 1);
+    variableReader.skipValues((long) index - 1L);
 
     final var offsetBefore = variableReader.getOffset();
     variableReader.skipValue();
